pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bid-crawler"
version = "0.1.0"
description = "Commercial construction bid aggregator for DFW public institutions"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "click>=8.1",
    "duckdb>=1.0",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
    "streamlit>=1.35",
    "plotly>=5.20",
    "pandas>=2.1",
    "pyarrow>=15.0",
    "lxml>=5.0",
]

[project.scripts]
bid-crawler = "bid_crawler.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["bid_crawler"]
